+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v2.20 - 2025-05-08
+
+- Remove database table to track schema
+    - Create new `migrate` command
+    - Build schema based on database structure
+
+- Change A,B,C modes to simpler Now / Later
+
+
 ## v2.1.3 - 2025-05-07
 
 - Added URL field to tasks #1